src/web-master/requirements.txt
gradio==5.49.1
pillow==11.3.0
uvicorn==0.38.0